chore: move outline to another file
